adrress comments, fix tests, fix si/ei
